v1.5.95 dev merge

+++ build-prod.sh
@@ -4,7 +4,11 @@
 set -e
 
 # clean build
-rm -rf build
+
+if [ "$1" != "--continue" ]; then
+    rm -rf build
+fi
+
 
 # configure for release
 mkdir -p build
